pyproject.toml
[tool.poetry]
name = "sbevebot"
version = "0.1.0"
description = ""
authors = ["Mohamed Ramdane DEBIANE <M.debiane71@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"discord.py" = { version = "*", extras = ["voice"] }
python-dotenv = "^1.0.0"
yt-dlp = "^2023.7.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
pytest = "^7.3.2"
pytest-cov = "^4.1.0"
flake8-pytest-style = "^1.7.2"
wemake-python-styleguide = "^0.18.0"
safety = "^2.3.5"
pre-commit = "^3.3.3"
nitpick = "^0.33.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
